Update tests to work around Sponge API's null keys.
